[refactor] mergeParams 위치 변경 및 [userId] 라우팅 생성
